Add residual connectionns.
I'm very confused.

This is meant to be analogous to this commit from Karpathy:

https://github.com/karpathy/ng-video-lecture/commit/5c3a2d299592603581129fdb14c5b06f5c50938c

However, when I used 3 blocks, I got terrible loss (about 3.0, which is
a regressionn from 2.27).

So, this keeps us at only 1 block, but still adds the residual
connection, which does seem to slightly improve the loss to 2.23.